% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotan_script.R
\name{find_closest_name}
\alias{find_closest_name}
\title{find_closest_name}
\usage{
find_closest_name(input, list_stations)
}
\arguments{
\item{input}{A string representing the input value.}

\item{list_stations}{A vector of strings representing station names.}
}
\value{
A string or NA, representing the name of the nearest station or NA if no close match is found.
}
\description{
The function takes two inputs, input and list_stations. It calculates the Levenshtein distance between the input and all elements of list_stations. The station with the smallest distance is considered the closest station and is returned. If the distance to the nearest station is 0, it means that the input is exactly equal to the station name and the station name is returned. If the distance is less than or equal to the length of the entry, the nearest station is returned. Otherwise, NA is returned.
}
