% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotan_script.R
\name{get_next_transport}
\alias{get_next_transport}
\title{get_next_transport at Nantes}
\usage{
get_next_transport(arret)
}
\arguments{
\item{input}{A character string representing the full station name, the "libelle"}
}
\value{
A data frame of the next passages for all lines passing through the station
}
\description{
This function allows you to obtain information on the schedules of the next trams
}
\examples{
get_next_transport("Bourgeonnière")
1      Orvault Grand Val   2mn
}
