% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotan_script.R
\name{get_waiting_time}
\alias{get_waiting_time}
\title{get_waiting_time}
\usage{
get_waiting_time(codeLieu)
}
\arguments{
\item{locationcode}{The station code for which the waiting time for the next trams is required}
}
\value{
A dataframe with information on the terminus and the time of the next trams
}
\description{
Retrieves information about the waiting time of the next trams
}
