% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotan_script.R
\name{is_station}
\alias{is_station}
\title{is_station}
\usage{
is_station(code)
}
\arguments{
\item{code}{Character string representing the station code.}
}
\value{
A logical value indicating whether the station exists.
}
\description{
The function takes an alphabetic code as input and returns a logical value indicating whether the station exists.
}
\examples{
is_station("ABCH")
is_station("MONERO")
}
